[moe training] split fp8 vs mxfp8 code into different files; refactor scaling_type -> recipe
stack-info: PR: https://github.com/pytorch/ao/pull/3852, branch: danielvegamyhre/stack/138
