+++ torchao/prototype/moe_training/__init__.py
@@ -1,9 +1,11 @@
-from torchao.prototype.moe_training.scaled_grouped_mm import (
+from torchao.prototype.moe_training.mxfp8_grouped_mm import (
     _quantize_then_scaled_grouped_mm,
+    _to_fp8_rowwise_then_scaled_grouped_mm,
     _to_mxfp8_then_scaled_grouped_mm,
 )
 
 __all__ = [
     "_quantize_then_scaled_grouped_mm",
     "_to_mxfp8_then_scaled_grouped_mm",
+    "_to_fp8_rowwise_then_scaled_grouped_mm",
 ]
