+++ torchao/prototype/moe_training/fp8_grouped_mm.py
@@ -1,0 +1,224 @@
+# Copyright (c) Meta Platforms, Inc. and affiliates.
+# All rights reserved.
+#
+# This source code is licensed under the BSD 3-Clause license found in the
+# LICENSE file in the root directory of this source tree.
+
+from typing import Optional
+
+import torch
+
+from torchao.float8.config import ScalingGranularity
+from torchao.float8.float8_utils import tensor_to_scale, to_fp8_saturated
+from torchao.prototype.moe_training.kernels import (
+    triton_fp8_per_group_colwise_scales,
+    triton_fp8_rowwise_3d_transpose_rhs,
+)
+from torchao.prototype.moe_training.utils import _is_column_major
+
+
+class _Float8GroupedMM(torch.autograd.Function):
+    """Differentiable implementation of grouped GEMM with dynamic float8 quantization."""
+
+    @staticmethod
+    def forward(
+        ctx,
+        A: torch.Tensor,
+        B_t: torch.Tensor,
+        offs: Optional[torch.Tensor] = None,
+        out_dtype: Optional[torch.dtype] = torch.bfloat16,
+    ) -> torch.Tensor:
+        # torchao _quantize_then_scaled_grouped_mm only supports A=2D|3D and B=3D.
+        assert A.ndim == 2 or A.ndim == 3, "A must be 2D or 3D"
+        assert B_t.ndim == 3, "B must be 3D"
+
+        assert A.size(-1) % 16 == 0, (
+            f"A must have a last dim divisible by 16, but got shape: {A.shape}"
+        )
+        assert B_t.size(-2) % 16 == 0 and B_t.size(-1) % 16 == 0, (
+            f"B must have last 2 dims divisible by 16, but got shape: {B_t.shape}"
+        )
+
+        # Assert input tensors are in high-precision dtypes.
+        assert A.dtype == torch.float32 or A.dtype == torch.bfloat16, (
+            "A must be float32 or bfloat16"
+        )
+        assert B_t.dtype == torch.float32 or B_t.dtype == torch.bfloat16, (
+            "B must be float32 or bfloat16"
+        )
+        assert offs is None or offs.dtype == torch.int32, (
+            "offs must be int32 tensor or None"
+        )
+
+        # Assert A and B dims are compatible for a scaled grouped GEMM.
+        assert A.size(-1) == B_t.size(-2), (
+            f"shape {A.shape} and {B_t.shape} are not compatible for _quantize_then_scaled_grouped_mm"
+        )
+
+        # The left operand in the scaled grouped GEMM must be row-major due to hardware requirements.
+        assert not _is_column_major(A), "A must be row-major"
+
+        # Due to hardware requirements, the right operand in a scaled grouped GEMM must be column-major.
+        assert _is_column_major(B_t), "B must be column-major"
+
+        # Convert high precision input tensor to float8, row-major for left operand of grouped GEMM.
+        # A shape: (M, K) or (B, M, K)
+        # A_scales shape: (M,1) or (B, M, 1)
+        A_scales = tensor_to_scale(
+            A,
+            torch.float8_e4m3fn,
+            scaling_granularity=ScalingGranularity.AXISWISE,
+            axiswise_dim=-1,
+            round_scales_to_power_of_2=True,
+        )
+        A_scaled = A.to(torch.float32) * A_scales
+        A_data_row_major = to_fp8_saturated(A_scaled, torch.float8_e4m3fn)
+
+        # Convert B to float8, column-major for right operand of grouped GEMM.
+        # B_t shape: (E, K, N)
+        # B_t scales must be computed rowwise keeping the outer/final dim, so:
+        # B_t_scales shape: (E, 1, N)
+        B_t_scales = tensor_to_scale(
+            B_t,
+            torch.float8_e4m3fn,
+            scaling_granularity=ScalingGranularity.AXISWISE,
+            axiswise_dim=-2,
+            round_scales_to_power_of_2=True,
+        )
+        B_t_scaled = B_t.to(torch.float32) * B_t_scales
+        B_t_data_col_major = to_fp8_saturated(B_t_scaled, torch.float8_e4m3fn)
+
+        # Store what we need for backward.
+        ctx.save_for_backward(A, B_t, offs)
+        ctx.out_dtype = out_dtype
+
+        # Perform scaled grouped GEMM and return result.
+        # output shape: scaled grouped mm of (M,K) @ (B,K,N) = (M,N)
+        assert not _is_column_major(A_data_row_major), (
+            "A must be row-major for output = A @ B"
+        )
+        assert _is_column_major(B_t_data_col_major), (
+            "B must be column-major for output = A @ B"
+        )
+
+        # Squeeze empty dims out of scales, to comply with grouped mm API.
+        # A_scales shape: (M,1) or (B, M, 1)
+        # B_t_scales shape: (E, 1, N)
+        A_scales = A_scales.squeeze(-1)
+        B_t_scales = B_t_scales.squeeze(1)
+        return torch._scaled_grouped_mm(
+            A_data_row_major,
+            B_t_data_col_major,
+            A_scales.reciprocal(),  # Reciprocals are needed for rescaling the output.
+            B_t_scales.reciprocal(),
+            offs,
+            out_dtype=out_dtype,
+            use_fast_accum=True,
+        )
+
+    @staticmethod
+    def backward(ctx, grad_output: torch.Tensor):
+        A, B_t, offs = ctx.saved_tensors
+        out_dtype = ctx.out_dtype
+
+        # Convert grad_output to float8, row-major for left operand of grouped GEMM
+        # needed for grad_A: grad_output @ B
+        #
+        # grad_output shape: (Mg, N)
+        # grad_output_scale shape: (Mg, 1)
+        grad_output_scales = tensor_to_scale(
+            grad_output,
+            torch.float8_e4m3fn,
+            scaling_granularity=ScalingGranularity.AXISWISE,
+            axiswise_dim=-1,
+            round_scales_to_power_of_2=True,
+        )
+        grad_output_scaled = grad_output.to(torch.float32) * grad_output_scales
+        grad_output_data_row_major = to_fp8_saturated(
+            grad_output_scaled, torch.float8_e4m3fn
+        )
+
+        # Compute B fp8 column-major for right operand of grouped GEMM:
+        # grad_A = grad_output @ B.
+        B_data_col_major, B_scales = triton_fp8_rowwise_3d_transpose_rhs(
+            B_t._data if hasattr(B_t, "_data") else B_t,
+            output_dtype=torch.float8_e4m3fn,
+            round_scales_to_power_of_2=True,
+        )
+
+        # Compute grad_A.
+        # grad_A = grad_output @ B
+        # grad_A = scaled grouped mm of (M,N) @ (B,N,K) = (M,K)
+        assert not _is_column_major(grad_output_data_row_major), (
+            "grad_output must be row-major for grad_A = grad_output @ B"
+        )
+        assert _is_column_major(B_data_col_major), (
+            "B must be column-major for grad_A = grad_output @ B"
+        )
+
+        # Squeeze empty dims out of scales, to comply with grouped mm API.
+        # grad_output_scales shape: (M,1) or (B, M, 1)
+        # B_scales shape: (E, 1, N)
+        grad_output_scales = grad_output_scales.squeeze(-1)
+        B_scales = B_scales.squeeze(1)
+        grad_A = torch._scaled_grouped_mm(
+            grad_output_data_row_major,
+            B_data_col_major,
+            grad_output_scales.reciprocal(),
+            B_scales.reciprocal(),
+            offs,
+            out_dtype=out_dtype,
+            use_fast_accum=True,
+        )
+
+        # grad_B is a special case. both operands of the grouped gemm will be 2D with offsets determing the "groups."
+        # Compute scales for grad_output_t and A, which are both 2D tensors with offsets which define the "jagged" groups.
+
+        # Convert transpose of grad_output to float8, row-major for left operand of grouped GEMM
+        # needed for grad_B: grad_output_t @ A
+        # Use transpose method to avoid uncoalesced memory accesses.
+        grad_out_data_colwise, grad_out_scales = triton_fp8_per_group_colwise_scales(
+            grad_output.t()
+            .contiguous()
+            .t(),  # Quantization is over 2x faster when input is col major, even with this transformation
+            offs,
+            torch.float8_e4m3fn,
+            round_scales_to_power_of_2=True,
+        )
+        grad_output_t_data_row_major = grad_out_data_colwise.t()
+        grad_output_t_scales = grad_out_scales.t()
+
+        A_data_col_major, A_scales = triton_fp8_per_group_colwise_scales(
+            A.t()
+            .contiguous()
+            .t(),  # Quantization is over 2x faster when input is col major, even with this transformation
+            offs,
+            torch.float8_e4m3fn,
+            round_scales_to_power_of_2=True,
+        )
+
+        # Compute grad_B = grad_output_t @ A.
+        # grad_B = grad_output_t @ A
+        assert not _is_column_major(grad_output_t_data_row_major), (
+            "grad_output_t must be row-major for grad_B = grad_output_t @ A"
+        )
+        assert _is_column_major(A_data_col_major), (
+            "A must be column-major for grad_B = grad_output_t @ A"
+        )
+
+        # Per-token group scales computed via triton kernels above do not have
+        # the empty dim like the scales computed via tensor_to_scale, so we need
+        # don't need to squeeze here.
+        grad_B = torch._scaled_grouped_mm(
+            grad_output_t_data_row_major,
+            A_data_col_major,
+            grad_output_t_scales.reciprocal(),
+            A_scales.reciprocal(),
+            offs,
+            out_dtype=out_dtype,
+            use_fast_accum=True,
+        )
+        return grad_A, grad_B.transpose(-2, -1), None, None, None, None
+
+
+_to_fp8_rowwise_then_scaled_grouped_mm = _Float8GroupedMM.apply
